Engine: modify user to be extension of core ::User
