+++ app/models/katello/activation_key.rb
@@ -18,7 +18,7 @@ class ActivationKey < ActiveRecord::Base
 
   belongs_to :organization
   belongs_to :environment, :class_name => "KTEnvironment"
-  belongs_to :user
+  belongs_to :user, :class_name => "::User"
   belongs_to :content_view, :inverse_of => :activation_keys
 
   has_many :key_pools, :dependent => :destroy
