+++ app/models/katello/user_own_role.rb
@@ -13,7 +13,7 @@
 module Katello
 class UserOwnRole < Role
 
-  use_index_of Role  if Katello.config.use_elasticsearch
+  use_index_of Katello::Role if Katello.config.use_elasticsearch
 
   def self_role_for_user
     users.first
